pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ops_automation"
version = "0.1.0"
description = "Wintel ops automation using Azure Arc, Defender for Cloud, and SRE Agent"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.17.0",
    "azure-mgmt-resource>=23.0.0",
    "azure-mgmt-compute>=32.0.0",
    "azure-monitor-query>=1.3.0",
    "azure-keyvault-secrets>=4.8.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
